Consolidate static constructor implementations
